This function performs a comprehensive analysis of initial and final datasets, including summary of differences, grouping differences, dimension comparisons, temporal and spatial analyses.

This function performs a comprehensive analysis of initial and final datasets, including summary of differences, grouping differences, dimension comparisons, temporal and spatial analyses.

Usage

comprehensive_cwp_dataframe_analysis(
  parameter_init,
  parameter_final,
  fig.path = getwd(),
  parameter_fact = "catch",
  parameter_short = FALSE,
  parameter_columns_to_keep = c("Precision", "measurement_unit", "Values dataset 1",
    "Values dataset 2", "Loss / Gain", "Difference (in %)", "Dimension",
    "Difference in value"),
  parameter_diff_value_or_percent = "Difference (in %)",
  parameter_UNK_for_not_standards_unit = TRUE,
  parameter_filtering = list(species = NULL, fishing_fleet = NULL),
  parameter_time_dimension = c("time_start"),
  parameter_geographical_dimension = "geographic_identifier",
  parameter_geographical_dimension_groupping = "GRIDTYPE",
  parameter_colnames_to_keep = "all",
  outputonly = FALSE,
  plotting_type = "view",
  print_map = TRUE,
  shapefile_fix = NULL,
  continent = NULL,
  coverage = TRUE,
  parameter_resolution_filter = NULL,
  parameter_titre_dataset_1 = "Dataset 1",
  parameter_titre_dataset_2 = "Dataset 2",
  unique_analyse = FALSE,
  removemap = FALSE,
  topnumber = 6
)

Arguments

fig.path

Path to save figures.

parameter_fact

Fact parameter, default is "#catch".

parameter_short

Whether to use short parameter names, default is FALSE.

parameter_columns_to_keep

Columns to keep in the comparison.

parameter_diff_value_or_percent

Difference value or percent, default is "Difference (in %)".

parameter_UNK_for_not_standards_unit

Whether to use UNK for non-standard units, default is TRUE.

parameter_filtering

Filtering parameters.

parameter_time_dimension

Time dimension parameters, default is c("time_start").

parameter_geographical_dimension

Geographical dimension parameter, default is "geographic_identifier".

parameter_geographical_dimension_groupping

Geographical dimension grouping parameter, default is "GRIDTYPE".

parameter_colnames_to_keep

Column names to keep in the parameter.

outputonly

Whether to output only.

print_map

Whether to print the map, default is FALSE.

coverage

Whether to analysis the time, geo and other dimension (if no analysing only summary)

parameter_resolution_filter

Resolution filter parameter.

parameter_titre_dataset_1

Title for dataset 1.

parameter_titre_dataset_2

Title for dataset 2.

unique_analyse

Whether the analysis is unique.

topnumber

Number of first caracteristics of each dimensions to display without groupping

init

Initial dataset.

final

Final dataset.

Value

A list containing results of various analyses.