Calculate and Visualize Data Distribution for Various Dimensions Using dygraphs

This function calculates and visualizes the data distribution for various dimensions using interactive time-series plots with dygraphs for two datasets or a single dataset.

Usage

other_dimension_analysis_dygraphs(
  Other_dimensions,
  init,
  final = NULL,
  titre_1 = "Dataset1",
  titre_2 = "Dataset2",
  unique_analyse = FALSE,
  fig.path = NULL,
  topn = 7
)

Arguments

Other_dimensions

A vector of dimensions to analyze.

init

Initial dataset.

final

Final dataset (optional).

titre_1

Title for the first dataset.

titre_2

Title for the second dataset (optional).

unique_analyse

Logical indicating whether the analysis is unique.

fig.path

Path to save the figures (optional).

topn

The number of top categories to include in the chart.

Value

A list containing the dygraphs for each dimension.

Examples

if (FALSE) { # \dontrun{
data_dimension_analysis_dygraphs(c("Dimension1", "Dimension2"), init, final, "Dataset1", "Dataset2", FALSE)
} # }