Read Data from Various File Types

This function reads data based on the provided file path and file type.

read_data(file_path)

Arguments

file_path

A character string indicating the path of the file to read.

Value

Data read from the specified file, either as a data frame or data table.

Examples

# Example: Reading a CSV file
csv_data <- read_data(here::here("inst/CWP_dataset.csv"))
head(csv_data)  # Display the first few rows
#>         long      lat unique_squirrel_id hectare  shift     date
#>        <num>    <num>             <char>  <char> <char>    <int>
#> 1: -73.96767 40.77647     15G-AM-1008-03     15G     AM 10082018
#> 2: -73.97603 40.77036      5D-AM-1020-02     05D     AM 10202018
#> 3: -73.95460 40.79566     39F-AM-1014-04     39F     AM 10142018
#> 4: -73.97159 40.77462     11E-PM-1012-06     11E     PM 10122018
#> 5: -73.95553 40.79777     40D-AM-1013-04     40D     AM 10132018
#> 6: -73.96479 40.79116     30A-PM-1020-02     30A     PM 10202018
#>    hectare_squirrel_number      age primary_fur_color highlight_fur_color
#>                      <int>   <char>            <char>              <char>
#> 1:                       3    Adult              Gray                <NA>
#> 2:                       2    Adult              Gray            Cinnamon
#> 3:                       4 Juvenile              Gray            Cinnamon
#> 4:                       6    Adult              Gray     Cinnamon, White
#> 5:                       4    Adult              Gray            Cinnamon
#> 6:                       2    Adult              Gray     Cinnamon, White
#>    combination_of_primary_and_highlight_color
#>                                        <char>
#> 1:                                      Gray+
#> 2:                              Gray+Cinnamon
#> 3:                              Gray+Cinnamon
#> 4:                       Gray+Cinnamon, White
#> 5:                              Gray+Cinnamon
#> 6:                       Gray+Cinnamon, White
#>                                                                                                      color_notes
#>                                                                                                           <char>
#> 1:                                                                                                          <NA>
#> 2:                                                                                                          <NA>
#> 3:                                                                                                          <NA>
#> 4:                Gray & Cinnamon selected as Primary. White selected as Highlights. Made executive adjustments.
#> 5:              Gray & Cinnamon selected as Primary. Nothing selected as Highlights. Made executive adjustments.
#> 6: Cinnamon back. Gray & Cinnamon selected as Primary. White selected as Highlights. Made executive adjustments.
#>        location above_ground_sighter_measurement specific_location running
#>          <char>                           <char>            <char>  <lgcl>
#> 1: Ground Plane                            FALSE              <NA>   FALSE
#> 2: Ground Plane                            FALSE              <NA>   FALSE
#> 3: Ground Plane                            FALSE              <NA>   FALSE
#> 4: Above Ground                                5              <NA>   FALSE
#> 5: Ground Plane                            FALSE              <NA>   FALSE
#> 6: Ground Plane                            FALSE              Rock   FALSE
#>    chasing climbing eating foraging other_activities   kuks  quaas  moans
#>     <lgcl>   <lgcl> <lgcl>   <lgcl>           <char> <lgcl> <lgcl> <lgcl>
#> 1:    TRUE    FALSE  FALSE    FALSE             <NA>   TRUE  FALSE  FALSE
#> 2:   FALSE     TRUE  FALSE     TRUE      small jumps  FALSE  FALSE  FALSE
#> 3:   FALSE    FALSE  FALSE     TRUE             <NA>  FALSE  FALSE  FALSE
#> 4:   FALSE     TRUE   TRUE    FALSE             <NA>  FALSE  FALSE  FALSE
#> 5:   FALSE    FALSE   TRUE     TRUE             <NA>  FALSE  FALSE  FALSE
#> 6:   FALSE    FALSE   TRUE    FALSE             <NA>  FALSE  FALSE  FALSE
#>    tail_flags tail_twitches approaches indifferent runs_from other_interactions
#>        <lgcl>        <lgcl>     <lgcl>      <lgcl>    <lgcl>             <char>
#> 1:      FALSE         FALSE      FALSE       FALSE     FALSE               <NA>
#> 2:      FALSE         FALSE      FALSE       FALSE     FALSE               <NA>
#> 3:      FALSE         FALSE      FALSE        TRUE     FALSE               <NA>
#> 4:      FALSE         FALSE      FALSE        TRUE     FALSE               <NA>
#> 5:      FALSE         FALSE      FALSE        TRUE     FALSE               <NA>
#> 6:      FALSE         FALSE      FALSE       FALSE      TRUE               <NA>
#>                                      lat_long zip_codes community_districts
#>                                        <char>     <int>               <int>
#> 1: POINT (-73.9676692772518 40.7764726688908)        NA                  19
#> 2: POINT (-73.9760256477558 40.7703573131398)        NA                  19
#> 3:  POINT (-73.954595058342 40.7956567934304)        NA                  19
#> 4: POINT (-73.9715921982161 40.7746153066601)        NA                  19
#> 5: POINT (-73.9555255950648 40.7977720808005)        NA                  19
#> 6: POINT (-73.9647904085226 40.7911589002217)        NA                  19
#>    borough_boundaries city_council_districts police_precincts
#>                 <int>                  <int>            <int>
#> 1:                  4                     19               13
#> 2:                  4                     19               13
#> 3:                  4                     19               13
#> 4:                  4                     19               13
#> 5:                  4                     19               13
#> 6:                  4                     19               13