Launch Strata in Georef but not in Nominal Report

This function launches a report comparing strata in georef data but not in nominal data.

Usage

strata_in_georef_but_not_in_nominal_report_launching(
  main.dir,
  connectionDB,
  uploadgoogledrive = TRUE
)

Arguments

main.dir

Directory containing the main data files.

connectionDB

Database connection object.

uploadgoogledrive

Logical indicating whether to upload the report to Google Drive.

Value

Data frame containing the upgraded nominal data.