Summarising_step

This function performs various summarizing steps on data related to species and gear types, retrieving data from a database, processing it, and rendering output reports.

Usage

summarising_step(
  main_dir,
  connectionDB,
  config,
  source_authoritylist = c("all", "IOTC", "WCPFC", "IATTC", "ICCAT", "CCSBT"),
  sizepdf = "long",
  savestep = FALSE,
  nameoutput = NULL,
  usesave = FALSE,
  fast_and_heavy = TRUE
)

Arguments

main_dir

Character. The main directory containing the entities. (jobs/entities)

connectionDB

Object. The database connection.

config

List. Configuration list containing metadata and options for processing.

source_authoritylist

Vector. Vector of source_authority to filter on, "all" being all of them.

sizepdf

Character string. La taille peut prendre les valeurs suivantes :

savestep

Logical TRUE/FALSE, should the .qs result of this be saved ?

nameoutput

Character, name of the .qs if saved

usesave

Logical Should we use the nameoutput .qs instead of rerunning everything ?

fast_and_heavy

Logical TRUE/FALSE, should we save a .qs for each dataset, and use this .qs in every markdown

"long" (par défaut) : Long with coverage.

"middle" : Long without coverage

"short" : Only first characteristics, first differences and main table of steps

Value

NULL. The function has side effects, such as writing files and rendering reports.

Examples

if (FALSE) { # \dontrun{
connectionDB <- DBI::dbConnect(RSQLite::SQLite(), ":memory:") # Connexion temporaire
config <- list() # Simule une configuration
summarising_step(main_dir = "chemin/vers/dossier", connectionDB, config)
} # }